requirements.txt
openai>=1.0.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
pdfplumber>=0.10.0
tiktoken>=0.5.0
presidio-analyzer>=2.2.0
presidio-anonymizer>=2.2.0
python-dotenv>=1.0.0
pyyaml>=6.0.0